+++ src/panels/lovelace/editor/dashboard-strategy-editor/dialogs/dialog-dashboard-strategy-editor.ts
@@ -6,22 +6,22 @@ import {
 } from "@mdi/js";
 import { CSSResultGroup, LitElement, css, html, nothing } from "lit";
 import { customElement, property, query, state } from "lit/decorators";
-import { HASSDomEvent, fireEvent } from "../../../../common/dom/fire_event";
-import { stopPropagation } from "../../../../common/dom/stop_propagation";
-import "../../../../components/ha-button";
-import "../../../../components/ha-button-menu";
-import "../../../../components/ha-dialog";
-import "../../../../components/ha-dialog-header";
-import "../../../../components/ha-icon-button";
-import { LovelaceStrategyConfig } from "../../../../data/lovelace/config/strategy";
-import { haStyleDialog } from "../../../../resources/styles";
-import type { HomeAssistant } from "../../../../types";
-import { showSaveSuccessToast } from "../../../../util/toast-saved-success";
-import "../../editor/dashboard-strategy-editor/hui-dashboard-strategy-element-editor";
-import type { HuiDashboardStrategyElementEditor } from "../../editor/dashboard-strategy-editor/hui-dashboard-strategy-element-editor";
-import { ConfigChangedEvent } from "../../editor/hui-element-editor";
-import { GUIModeChangedEvent } from "../../editor/types";
-import { cleanLegacyStrategyConfig } from "../legacy-strategy";
+import { HASSDomEvent, fireEvent } from "../../../../../common/dom/fire_event";
+import { stopPropagation } from "../../../../../common/dom/stop_propagation";
+import "../../../../../components/ha-button";
+import "../../../../../components/ha-button-menu";
+import "../../../../../components/ha-dialog";
+import "../../../../../components/ha-dialog-header";
+import "../../../../../components/ha-icon-button";
+import { LovelaceStrategyConfig } from "../../../../../data/lovelace/config/strategy";
+import { haStyleDialog } from "../../../../../resources/styles";
+import type { HomeAssistant } from "../../../../../types";
+import { showSaveSuccessToast } from "../../../../../util/toast-saved-success";
+import "../hui-dashboard-strategy-element-editor";
+import type { HuiDashboardStrategyElementEditor } from "../hui-dashboard-strategy-element-editor";
+import { ConfigChangedEvent } from "../../hui-element-editor";
+import { GUIModeChangedEvent } from "../../types";
+import { cleanLegacyStrategyConfig } from "../../../strategies/legacy-strategy";
 import type { DashboardStrategyEditorDialogParams } from "./show-dialog-dashboard-strategy-editor";
 
 @customElement("dialog-dashboard-strategy-editor")
