+++ src/panels/lovelace/strategies/original-states/original-states-view-strategy.ts
@@ -1,12 +1,12 @@
 import { STATE_NOT_RUNNING } from "home-assistant-js-websocket";
 import { ReactiveElement } from "lit";
 import { customElement } from "lit/decorators";
-import { isComponentLoaded } from "../../../common/config/is_component_loaded";
-import type { AreaFilterValue } from "../../../components/ha-area-filter";
-import { getEnergyPreferences } from "../../../data/energy";
-import { LovelaceViewConfig } from "../../../data/lovelace/config/view";
-import { HomeAssistant } from "../../../types";
-import { generateDefaultViewConfig } from "../common/generate-lovelace-config";
+import { isComponentLoaded } from "../../../../common/config/is_component_loaded";
+import type { AreaFilterValue } from "../../../../components/ha-area-filter";
+import { getEnergyPreferences } from "../../../../data/energy";
+import { LovelaceViewConfig } from "../../../../data/lovelace/config/view";
+import { HomeAssistant } from "../../../../types";
+import { generateDefaultViewConfig } from "../../common/generate-lovelace-config";
 
 export type OriginalStatesViewStrategyConfig = {
   type: "original-states";
